+++ src/enums/node-transformers/NodeTransformer.ts
@@ -1,6 +1,7 @@
 export enum NodeTransformer {
     BlockStatementControlFlowTransformer,
     ClassDeclarationTransformer,
+    CommentsTransformer,
     DeadCodeInjectionTransformer,
     FunctionControlFlowTransformer,
     CatchClauseTransformer,