+++ test/functional-tests/node-transformers/preparing-transformers/comments-transformer/fixtures/combined-words-1.js
@@ -1,0 +1,3 @@
+// test comment
+// @license test comment
+var test = 1;