+++ test/functional-tests/node-transformers/preparing-transformers/comments-transformer/fixtures/no-preserved-words.js
@@ -1,0 +1,2 @@
+// test comment
+var test = 1;